test(compose): the review round's corrections to the agent-access slice
Six findings, and the first is the one that mattered most because it lands in the
durable record: STATUS.md credited this slice to #877, which is a different,
already-merged PR (the license gate). Corrected to #913 in all three places. A
recorded reference that does not resolve is worse than none.

The doc.go boundary rationale was wrong, not merely thin. It said agentscope and
approval_bundle stayed because they 'turn on approvals, not on admission' — but
agentscope's own header says it proves the passport cap over real HTTP, which is
admission verbatim, and oauth_surface moved despite turning on an approval token.
The criterion was applied in both directions. The real reason is the one channels
gives for its neighbours: the assertions they need (assertNothingStaged, the
approval-inbox queries) live in the parent's _test.go files, which a subpackage
cannot reach. A wrong reason is worse than a missing one here, because the next
slice will apply it.

AppDSN moved from suitefixtures.go to apptest, where it should have gone. It was a
third copy of one invariant: apptest/earlypool.go read the same env var with the
same fatal string, and apptest cannot import integration, so a copy in
suitefixtures.go permanently excluded apptest's own callers. In apptest it serves
every caller, and earlypool.go now calls it — one spelling instead of three.

The inlined custom-field create had dropped the refusal body from its failure
message, in the one case the comment above it names as most likely (501 with no
schema pool), where the detail IS the diagnosis. It decodes title/detail again.

Two comments repaired: readPassport justified its shape against 'the REST helper
next to it', which the promotion moved to another package, and PassportBearer
hedged its rationale as what 'most callers' prove when the invariant is
unconditional. doc.go also no longer claims the suites share a minted-passport
fixture; two of them use it, the rest mint their own for different shapes.
