+++ backend/internal/compose/integration/apptest/earlypool.go
@@ -7,7 +7,6 @@ package apptest
 
 import (
 	"context"
-	"os"
 	"testing"
 
 	"github.com/jackc/pgx/v5/pgxpool"
@@ -25,11 +24,7 @@ import (
 // *testing.T has no reason to sit in either one of them.
 func EarlyPool(t *testing.T) *pgxpool.Pool {
 	t.Helper()
-	appDSN := os.Getenv("MARGINCE_TEST_APP_DSN")
-	if appDSN == "" {
-		t.Fatal("MARGINCE_TEST_APP_DSN not set — run `make db-up` (integration tests fail loudly, they never skip)")
-	}
-	pool, err := database.NewPool(context.Background(), appDSN)
+	pool, err := database.NewPool(context.Background(), AppDSN(t))
 	if err != nil {
 		t.Fatalf("opening the pre-boot pool: %v", err)
 	}
