+++ backend/internal/compose/integration/apptest/passport.go
@@ -13,13 +13,13 @@ import (
 // PassportBearer mints a passport with exactly scopes and returns the
 // Authorization header an agent principal presents.
 //
-// Minted through the real endpoint rather than written into the table, because
-// what most callers are proving is that the granting human's live seat and RBAC
-// still cap the passport — a hand-inserted row would skip the admission that
-// caps it.
+// Minted through the real endpoint, never written into the table: the granting
+// human's live seat and RBAC cap a passport at the moment of issue, so a
+// hand-inserted row is a credential production would not have issued and proves
+// nothing about one that was.
 //
-// It lives here because it is keyed on AppEnv and suites on both sides of the
-// integration/agentaccess boundary present passports.
+// It lives here because it is keyed on AppEnv and more than one suite package
+// presents passports.
 func PassportBearer(t *testing.T, e *AppEnv, label string, scopes ...string) map[string]string {
 	t.Helper()
 	var minted struct {
