+++ backend/internal/compose/integration/humanprecedence_integration_test.go
@@ -180,7 +180,7 @@ func stagePersonAndAgent(t *testing.T, e *apptest.AppEnv, fullName, label string
 // passport per call, exactly as the stdio/hosted transports do.
 func mcpAgentInvoker(t *testing.T, agentToken string) func(tool, args string) (json.RawMessage, error) {
 	t.Helper()
-	pool, err := database.NewPool(context.Background(), AppDSN(t))
+	pool, err := database.NewPool(context.Background(), apptest.AppDSN(t))
 	if err != nil {
 		t.Fatal(err)
 	}
